根据请求URL生成邮件中的链接；限制删除当前激活的组织；客户端导航栏刷新删除后的组织列表；

+++ src/main/java/com/ngtesting/platform/service/impl/OrgServiceImpl.java
@@ -64,9 +64,12 @@ public class OrgServiceImpl extends BaseServiceImpl implements OrgService {
 	}
 
 	@Override
-	public Boolean delete(Integer id) {
-		Integer count = orgDao.delete(id);
+	public Boolean delete(Integer id, TstUser user) {
+        if (user.getDefaultOrgId().intValue() == id.intValue()) {
+            return false;
+        }
 
+		Integer count = orgDao.delete(id);
 		return count > 0;
 	}
